revamped keras interface into an initial client
