+++ grid/__init__.py
@@ -1,2 +1,3 @@
 from . import client
+from . import clients
 from .client import *
